DetailsRefactoring
