+++ App/Page/DetailPage.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import { View, Text, TextInput, Button, StyleSheet} from "react-native";
 import CalculateEquations from "../Components/DetailPage/CalculateEquation.js";
+import HomePageComponents from '../components/DetailPageComponents.js';
 
 export default class DetailPage extends React.Component {
 
@@ -21,44 +22,7 @@ export default class DetailPage extends React.Component {
   }
 
   render() {
-      let payments = [];
-      for(let i = 0; i < this.equation.parameters.length; i++){
-        payments.push(
-          <View key={i}>
-            <Text>{this.equation.parameters[i].var}</Text>
-            <TextInput
-              style={ [
-                styles.equation,
-                {borderColor: (!this.state.parametersValidation[i] ? 'gray' : 'red')}
-              ]}
-              onChangeText={(text) => this.onParametersInput(i, text)}
-              value={this.state.parameterArray[i]}
-              keyboardType={'numeric'}
-            />
-            {!!this.state.parametersValidation[i] && (
-              <Text style={styles.validationTxtBox}>{this.state.parametersValidation[i]}</Text>
-            )}
-          </View>)
-      }
   
-      return (
-        <View>
-          <Text>{this.equation.name}</Text>
-          <Text>{this.equation.description}</Text>
-          <Text>{this.equation.equation}</Text>
-  
-          <View style={styles.equationParameters}>
-            {payments}
-          </View>
-  
-          <Text>{this.state.calculateResult}</Text>
-  
-          <Button
-            onPress={ () => this.onCalculatePress()}
-            title="Calculate"
-          />
-        </View>
-      )
   }
 
   onParametersInput(index, text) {
@@ -115,7 +79,7 @@ export default class DetailPage extends React.Component {
     let calculateResult = ""
     calculateResult = c.Calculate(this.state.parameterArray, this.equation);
 
-    this.updateCalculateReulsts(eval(calculateResult))
+    this.updateCalculateReulsts(calculateResult)
   }
 
   updateCalculateReulsts(calculateResult) {
